# assets/docs/openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spallet",
    "description": "Split Balance Wallet API Documentation",
    "version": "0.1.0"
  },
  "paths": {
...
    "/wallets/{user_id}/balance": {
      "get": {
        "operationId": "balance",
        "tags": [
          "wallet"
        ],
        "summary": "Balance",
        "parameters": [
          {
            "$ref": "#/components/parameters/user_id"
          }
        ],
        "responses": {
          "200": {
            "description": "The request was successful, and the server has returned the requested resource in the response body.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Balance"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "title": "BalanceOk"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Bad Request"
          },
          "404": {
            "$ref": "#/components/responses/Not Found"
          }
        },
        "description": "Calculate and return the user's balance."
      }
    }
  },
  "components": {
...
    "parameters": {
      "user_id": {
        "name": "user_id",
        "required": true,
        "in": "path",
        "description": "User ID",
        "schema": {
          "type": "integer"
        },
        "explode": false,
        "style": "simple",
        "allowReserved": false,
        "allowEmptyValue": false
      }
    },
    "schemas": {
...
      "Balance": {
        "type": "object",
        "properties": {
          "available": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "available",
          "total"
        ],
        "description": "Account balance details"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:4004",
      "description": "Local development address."
    }
  ]
}